Skip download_ace.sh if ace is already downloaded

+++ addons/cb.files.editor/download_ace.sh
@@ -16,6 +16,13 @@ DEST="${DIR}/ace"
 # Make sure $DEST exists
 mkdir -p ${DEST}
 
+# Check if it's already downloaded
+if [ -f "${DEST}/ace.js" ]; then
+    echo "Ace is already downloaded"
+    echo "'rm -rf ace' if you want to redownload"
+    exit 0
+fi
+
 # Download tar.gz and pipe to tar
 # decompressing it to $DEST
 wget -O - ${ACE_URL} | tar -xzv -C ${DEST} --strip-components=2 ace-builds-${ACE_VERSION}/${ACE_SUB}
